+++ sdmx-dl-cli/src/main/java/internal/sdmxdl/cli/NetworkOptions.java
@@ -1,10 +1,12 @@
 package internal.sdmxdl.cli;
 
+import internal.http.curl.CurlHttpURLConnection;
 import internal.sdmxdl.cli.ext.AuthOptions;
 import internal.sdmxdl.cli.ext.CacheOptions;
 import internal.sdmxdl.cli.ext.ProxyOptions;
 import internal.sdmxdl.cli.ext.SslOptions;
 import picocli.CommandLine;
+import sdmxdl.web.URLConnectionFactory;
 
 @lombok.Getter
 @lombok.Setter
@@ -17,6 +19,12 @@ public class NetworkOptions {
     )
     private boolean dummyNetworkOption;
 
+    @CommandLine.Option(
+            names = "--curl-backend",
+            defaultValue = "false"
+    )
+    private boolean curlBackend;
+
     @CommandLine.ArgGroup(validate = false)
     private CacheOptions cacheOptions = new CacheOptions();
 
@@ -28,4 +36,8 @@ public class NetworkOptions {
 
     @CommandLine.ArgGroup(validate = false)
     private AuthOptions authOptions = new AuthOptions();
+
+    public URLConnectionFactory getURLConnectionFactory() {
+        return isCurlBackend() ? CurlHttpURLConnection::of : URLConnectionFactory.getDefault();
+    }
 }