+++ sdmx-dl-provider-util/src/test/java/internal/http/curl/CurlTest.java
@@ -1,7 +1,8 @@
-package internal.util.http.curl;
+package internal.http.curl;
 
-import internal.util.http.HttpHeadersBuilder;
 import nbbrd.io.Resource;
+import nbbrd.io.sys.ProcessReader;
+import org.junit.jupiter.api.Disabled;
 import org.junit.jupiter.api.Test;
 
 import java.io.BufferedReader;
@@ -12,6 +13,7 @@ import java.net.InetSocketAddress;
 import java.net.Proxy;
 import java.nio.charset.StandardCharsets;
 
+import static java.util.Collections.singletonList;
 import static org.assertj.core.api.Assertions.assertThat;
 
 public class CurlTest {
@@ -38,22 +40,29 @@ public class CurlTest {
                                 .builder()
                                 .code(200)
                                 .message("OK")
-                                .headers(new HttpHeadersBuilder()
-                                        .put("Date", "Wed, 20 Oct 2021 10:58:37 GMT")
-                                        .put("Expires", "-1")
-                                        .put("Cache-Control", "private, max-age=0")
-                                        .put("Content-Type", "text/html; charset=ISO-8859-1")
-                                        .put("P3P", "CP=\"This is not a P3P policy! See g.co/p3phelp for more info.\"")
-                                        .put("Server", "gws")
-                                        .put("X-XSS-Protection", "0")
-                                        .put("X-Frame-Options", "SAMEORIGIN")
-                                        .put("Accept-Ranges", "none")
-                                        .put("Vary", "Accept-Encoding")
-                                        .put("Transfer-Encoding", "chunked")
-                                        .build())
+                                .header("Date", singletonList("Wed, 20 Oct 2021 10:58:37 GMT"))
+                                .header("Expires", singletonList("-1"))
+                                .header("Cache-Control", singletonList("private, max-age=0"))
+                                .header("Content-Type", singletonList("text/html; charset=ISO-8859-1"))
+                                .header("P3P", singletonList("CP=\"This is not a P3P policy! See g.co/p3phelp for more info.\""))
+                                .header("Server", singletonList("gws"))
+                                .header("X-XSS-Protection", singletonList("0"))
+                                .header("X-Frame-Options", singletonList("SAMEORIGIN"))
+                                .header("Accept-Ranges", singletonList("none"))
+                                .header("Vary", singletonList("Accept-Encoding"))
+                                .header("Transfer-Encoding", singletonList("chunked"))
                                 .build()
                         );
             }
+        }
+    }
+
+    @Disabled
+    @Test
+    public void testVersion() throws IOException {
+        String[] versionCommand = new Curl.CurlCommandBuilder().version().build();
+        try (BufferedReader reader = ProcessReader.newReader(versionCommand)) {
+            Curl.CurlVersion.parseText(reader).getLines().forEach(System.out::println);
         }
     }
 }